src/main/java/se/lexicon/model/Person.java
package se.lexicon.model;

public class Person {
    //Creating fields
    private int personId;
    private String firstName;
    private String lastName;

    //Constructing the object
    public Person(int personId,String firstName, String lastName) {
        this.personId = personId;
        this.firstName = firstName;
        this.lastName = lastName;
    }

    //Including getters and setters that are needed
    public int getPersonId() {
        return personId;
    }

    public void setPersonId(int personId) {
        this.personId = personId;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }
}
